main/requirements.txt

aiogram==3.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.32.3
